Move SoundLevelPath builder to util and rename to GeoJsonBuilder
